refactor: update code plus javadoc
