--- core/flamingock-test-support/src/main/java/io/flamingock/internal/core/builder/AuditExpectations.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2025 Flamingock (https://www.flamingock.io)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.flamingock.internal.core.builder;
-
-import io.flamingock.internal.common.core.audit.AuditEntry;
-
-public final class AuditExpectations {
-
-    private AuditExpectations() {}
-
-    public static AuditEntryExpectation applied(Class<?> changeClass) {
-        return AuditEntryExpectation
-                .auditEntry()
-                .withClass(changeClass)
-                .withState(AuditEntry.Status.APPLIED);
-    }
-
-    public static AuditEntryExpectation failed(Class<?> changeClass) {
-        return AuditEntryExpectation
-                .auditEntry()
-                .withClass(changeClass)
-                .withState(AuditEntry.Status.FAILED);
-    }
-
-    public static AuditEntryExpectation started(Class<?> changeClass) {
-        return AuditEntryExpectation
-                .auditEntry()
-                .withClass(changeClass)
-                .withState(AuditEntry.Status.STARTED);
-    }
-
-    public static AuditEntryExpectation rolledBack(Class<?> changeClass) {
-        return AuditEntryExpectation
-                .auditEntry()
-                .withClass(changeClass)
-                .withState(AuditEntry.Status.ROLLED_BACK);
-    }
-
-    public static AuditEntryExpectation rollbackFailed(Class<?> changeClass) {
-        return AuditEntryExpectation
-                .auditEntry()
-                .withClass(changeClass)
-                .withState(AuditEntry.Status.ROLLBACK_FAILED);
-    }
-}
