+++ core/flamingock-test-support/src/main/java/io/flamingock/internal/core/builder/BuilderAccessor.java
@@ -1,0 +1,16 @@
+package io.flamingock.internal.core.builder;
+
+import io.flamingock.internal.core.store.AuditStore;
+
+public class BuilderAccessor {
+
+    private AbstractChangeRunnerBuilder<?, ?> builder;
+
+    public BuilderAccessor(AbstractChangeRunnerBuilder<?,?> builder) {
+        this.builder = builder;
+    }
+
+    public AuditStore<?> getAuditStore(){
+        return builder.auditStore;
+    }
+}
